fix: consolidate test fixtures and add missing mock methods
- Move tts_client fixture to conftest.py (single location)
- Add audio_to_wav_bytes and get_audio_length_seconds mock methods
- Remove duplicate fixtures from test files
- Fix ruff lint issues (unused imports, line length)
