+++ backend/tests/integration/test_tts_api.py
@@ -3,23 +3,7 @@
 These tests validate the full request/response flow through the API.
 """
 
-import pytest
-from httpx import ASGITransport, AsyncClient
-
-
-@pytest.fixture
-async def tts_client(mock_tts_service):
-    """Create an async test client with mocked TTS service."""
-    from src.dependencies import set_tts_service
-    from src.main import app
-
-    set_tts_service(mock_tts_service)
-
-    async with AsyncClient(
-        transport=ASGITransport(app=app),
-        base_url="http://test",
-    ) as ac:
-        yield ac
+from httpx import AsyncClient
 
 
 class TestSynthesizeEndpoint:
